+++ ios/Sources/AppFeature/AppFeature.swift
@@ -6,32 +6,20 @@ import MediaFeature
 import Model
 import Repository
 
-public struct AppState: Equatable {
-    public enum AppCoreState: Equatable {
-        case needToInitialize
-        case initialized(AppTabState)
-        case errorOccurred
-    }
+public enum AppState: Equatable {
+    case needToInitialize
+    case initialized(AppTabState)
+    case errorOccurred
 
-    public var feedContents: [FeedContent]
-    public var coreState: AppCoreState
-
-    public init(
-        feedContents: [FeedContent] = [],
-        coreState: AppCoreState = .needToInitialize
-    ) {
-        self.feedContents = feedContents
-        self.coreState = coreState
+    public init() {
+        self = .needToInitialize
     }
 }
 
 public enum AppAction {
     case refresh
     case needRefresh
-    case selectFeedContent
     case refreshResponse(Result<[FeedContent], KotlinError>)
-    case tapFavorite(isFavorited: Bool, id: String)
-    case favoriteResponse(Result<String, KotlinError>)
     case appTab(AppTabAction)
 }
 
@@ -42,31 +30,13 @@ public let appReducer = Reducer<AppState, AppAction, AppEnvironment> { state, ac
             .catchToEffect()
             .map(AppAction.refreshResponse)
     case .needRefresh:
-        state.coreState = .needToInitialize
+        state = .needToInitialize
         return .none
     case let .refreshResponse(.success(feedContents)):
-        state.coreState = .initialized(.init(feedContent: feedContents))
+        state = .initialized(.init(feedContents: feedContents))
         return .none
     case let .refreshResponse(.failure(error)):
-        state.coreState = .errorOccurred
-        return .none
-    case .selectFeedContent:
-        return .none
-    case .tapFavorite(let isFavorited, let id):
-        if let index = state.feedContents.map(\.id).firstIndex(of: id) {
-            state.feedContents[index].isFavorited.toggle()
-        }
-        let publisher = isFavorited
-            ? environment.feedRepository.removeFavorite(id: id)
-            : environment.feedRepository.addFavorite(id: id)
-        return publisher
-            .map { id }
-            .catchToEffect()
-            .map(AppAction.favoriteResponse)
-    case let .favoriteResponse(.success(id)):
-        return .none
-    case let .favoriteResponse(.failure(error)):
-        print(error.localizedDescription)
+        state = .errorOccurred
         return .none
     case .appTab:
         return .none
