+++ ios/Sources/AppFeature/AppTabScreen.swift
@@ -5,67 +5,9 @@ import FavoritesFeature
 import AboutFeature
 import SwiftUI
 import Model
+import Styleguide
 
-public struct AppTabState: Equatable {
-    public var homeState: HomeState
-    public var mediaState: MediaState
-    public var favoritesState: FavoritesState
-    public var aboutState: AboutState
-
-    public init(
-        feedContent: [FeedContent]
-    ) {
-        self.homeState = HomeState.init(feedContents: feedContent)
-        self.mediaState = MediaState.init()
-        self.favoritesState = FavoritesState.init()
-        self.aboutState = AboutState.init()
-    }
-}
-
-public enum AppTabAction {
-    case home(HomeAction)
-    case media(MediaAction)
-    case favorites(FavoritesAction)
-    case about(AboutAction)
-}
-
-public let appTabReducer = Reducer<AppTabState, AppTabAction, AppEnvironment>.combine(
-    mediaReducer.pullback(
-        state: \.mediaState,
-        action: /AppTabAction.media,
-        environment: { environment in
-            .init(feedRepository: environment.feedRepository)
-        }
-    ),
-    favoritesReducer.pullback(
-        state: \.favoritesState,
-        action: /AppTabAction.favorites,
-        environment: { environment -> FavoritesEnvironment in
-            .init(feedRepository: environment.feedRepository)
-        }
-    ),
-    aboutReducer.pullback(
-        state: \.aboutState,
-        action: /AppTabAction.about,
-        environment: { _ -> AboutEnvironment in
-            .init()
-        }
-    ),
-    .init { _, action, _ in
-        switch action {
-        case .home:
-            return .none
-        case .media:
-            return .none
-        case .favorites:
-            return .none
-        case .about:
-            return .none
-        }
-    }
-)
-
-public struct AppTabView: View {
+public struct AppTabScreen: View {
     @State var selection: Int = .zero
     public let store: Store<AppTabState, AppTabAction>
 
@@ -98,8 +40,62 @@ public struct AppTabView: View {
                 .default
                 .publisher(for: UIApplication.willEnterForegroundNotification)
         ) { _ in
-            print("hoge")
-//            viewStore.send(.reload)
+            ViewStore(store).send(.reload)
         }
     }
 }
+
+private extension AppTab {
+    var title: String {
+        switch self {
+        case .home:
+            return L10n.HomeScreen.title
+        case .media:
+            return L10n.MediaScreen.title
+        case .favorites:
+            return L10n.FavoriteScreen.title
+        case .about:
+            return L10n.AboutScreen.title
+        }
+    }
+
+    var image: SwiftUI.Image {
+        switch self {
+        case .home:
+            return AssetImage.iconHome.image
+        case .media:
+            return AssetImage.iconBlog.image
+        case .favorites:
+            return AssetImage.iconStar.image
+        case .about:
+            return AssetImage.iconAbout.image
+        }
+    }
+}
+
+#if DEBUG
+ public struct AppTabScreen_Previews: PreviewProvider {
+    public static var previews: some View {
+        ForEach(ColorScheme.allCases, id: \.self) { colorScheme in
+            AppTabScreen(
+                store: .init(
+                    initialState: .init(
+                        feedContents: [
+                            .blogMock(),
+                            .blogMock(),
+                            .blogMock(),
+                            .blogMock(),
+                            .blogMock(),
+                            .blogMock()
+                        ]
+                    ),
+                    reducer: .empty,
+                    environment: {}
+                )
+            )
+            .previewDevice(.init(rawValue: "iPhone 12"))
+            .environment(\.colorScheme, colorScheme)
+        }
+    }
+ }
+#endif
