+++ ios/Sources/FavoritesFeature/FavoritesFeature.swift
@@ -3,62 +3,28 @@ import ComposableArchitecture
 import Model
 import Repository
 
-public enum FavoritesState: Equatable {
-    case needToInitialize
-    case emptyInitialized
-    case initialized(FavoritesListState)
+public struct FavoritesState: Equatable {
+    public var feedContents: [FeedContent]
 
-    public init() {
-        self = .needToInitialize
+    public init(feedContents: [FeedContent]) {
+        self.feedContents = feedContents
     }
 }
 
 public enum FavoritesAction {
-    case refresh
-    case refreshResponse(Result<[FeedContent], KotlinError>)
-    case needRefresh
-    case favoritesList(FavoritesListAction)
+    case tap(FeedContent)
+    case tapFavorite(isFavorited: Bool, id: String)
 }
 
 public struct FavoritesEnvironment {
-    public let feedRepository: FeedRepositoryProtocol
-
-    public init(
-        feedRepository: FeedRepositoryProtocol
-    ) {
-        self.feedRepository = feedRepository
-    }
+    public init() {}
 }
 
-public let favoritesReducer = Reducer<FavoritesState, FavoritesAction, FavoritesEnvironment>.combine(
-    favoritesListReducer.pullback(
-        state: /FavoritesState.initialized,
-        action: /FavoritesAction.favoritesList,
-        environment: {
-            .init(feedRepository: $0.feedRepository)
-        }
-    ),
-    .init { state, action, environment in
-        switch action {
-        case .refresh:
-            return environment.feedRepository.feedContents()
-                .catchToEffect()
-                .map(FavoritesAction.refreshResponse)
-        case let .refreshResponse(.success(feedContents)):
-            let filteredFeedContents = feedContents.filter(\.isFavorited)
-            state = filteredFeedContents.isEmpty
-                ? .emptyInitialized
-                : .initialized(.init(feedContents: filteredFeedContents))
-            return .none
-        case let .refreshResponse(.failure(error)):
-            print(error.localizedDescription)
-            // TODO: Error handling
-            return .none
-        case .needRefresh:
-            state = .needToInitialize
-            return .none
-        case .favoritesList:
-            return .none
-        }
+public let favoritesReducer = Reducer<FavoritesState, FavoritesAction, FavoritesEnvironment> { _, action, _ in
+    switch action {
+    case .tap:
+        return .none
+    case .tapFavorite(isFavorited: let isFavorited, id: let id):
+        return .none
     }
-)
+}
